refactor(application): extract command creation logic to CommandBuilder class
