+++ .env.example
@@ -10,3 +10,4 @@ APP_DEBUG=false
 APP_LICENSE=MIT
 APP_SIGNING_KEY=
 SUDO_PASSWORD=
+DISABLE_COSMIC_COMMANDS=false
