+++ src/Command/Builder/Validator/Exception/InvalidCallableException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ninja\Cosmic\Command\Builder\Validator\Exception;
+
+use Ninja\Cosmic\Exception\CosmicException;
+
+class InvalidCallableException extends CosmicException {}
